# core-common-org-mgmt-sdk/src/main/resources/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-common-org-mgmt
  description: Organization Management Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/organization
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/branches/{branchId}:
    get:
      tags:
        - Branch Management
      summary: Get branch by ID
      description: Returns a branch based on its ID
      operationId: getBranchById
      parameters:
        - name: branchId
          in: path
          description: ID of the branch to retrieve
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successfully retrieved branch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchDTO'
        '404':
          description: Branch not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchDTO'
    put:
      tags:
        - Branch Management
      summary: Update branch
      description: Updates an existing branch with the provided details
      operationId: updateBranch
      parameters:
        - name: branchId
          in: path
          description: ID of the branch to update
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BranchDTO'
        required: true
      responses:
        '200':
          description: Branch successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchDTO'
        '400':
          description: Invalid branch data supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchDTO'
        '404':
          description: Branch not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchDTO'
    delete:
      tags:
        - Branch Management
      summary: Delete branch
      description: Deletes a branch based on its ID
      operationId: deleteBranch
      parameters:
        - name: branchId
          in: path
          description: ID of the branch to delete
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Branch successfully deleted
        '404':
          description: Branch not found
        '500':
          description: Internal server error
...
  /api/v1/branches/filter:
    post:
      tags:
        - Branch Management
      summary: Get all branches with filtering
      description: Returns a paginated list of branches based on filter criteria
      operationId: filterBranches
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FilterRequestBranchDTO'
        required: true
      responses:
        '200':
          description: Successfully retrieved branches
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Invalid filter criteria supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseBranchDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginationResponseBranchDTO'
  /api/v1/branches:
    post:
      tags:
        - Branch Management
      summary: Create a new branch
      description: Creates a new branch with the provided details
      operationId: createBranch
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BranchDTO'
        required: true
      responses:
        '201':
          description: Branch successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchDTO'
        '400':
          description: Invalid branch data supplied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchDTO'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BranchDTO'
...
components:
  schemas:
...
    BranchDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        bankId:
          type: string
          format: uuid
        regionId:
          type: string
          format: uuid
        code:
          type: string
        name:
          type: string
        description:
          type: string
        phoneNumber:
          type: string
        email:
          type: string
        addressLine:
          type: string
        postalCode:
          type: string
        city:
          type: string
        state:
          type: string
        countryId:
          type: string
          format: uuid
        timeZoneId:
          type: string
          format: uuid
        latitude:
          type: number
          format: float
        longitude:
          type: number
          format: float
        isActive:
          type: boolean
        openedAt:
          type: string
          format: date-time
        closedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
          format: uuid
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: string
          format: uuid
...
    FilterOptions:
      type: object
      properties:
        caseInsensitiveStrings:
          type: boolean
        includeInheritedFields:
          type: boolean
      description: Filter options for customizing filter behavior
...
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: The direction of sorting, either ASC or DESC.
          example: DESC
          default: DESC
      description: Represents a pagination request for retrieving paginated results, including page number, size, sort field, and direction.
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/RangeObject'
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
    RangeObject:
      type: object
      properties:
        from:
          type: object
          description: Start value of the range
        to:
          type: object
          description: End value of the range
      description: Represents a range with from and to values
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...
    FilterRequestBranchDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/BranchDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Filter criteria for branches
    PaginationResponseBranchDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/BranchDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...